explicitely cast getpid() to (int) before sprintf()'ing it with "%d"

+++ locks.c
@@ -1,4 +1,4 @@
-#ident "$Id: locks.c,v 3.5 1996/01/22 01:12:59 gert Exp $ Copyright (c) Gert Doering / Paul Sutcliffe Jr."
+#ident "$Id: locks.c,v 3.6 1996/06/18 22:53:04 gert Exp $ Copyright (c) Gert Doering / Paul Sutcliffe Jr."
 
 /* large parts of the code in this module are taken from the
  * "getty kit 2.0" by Paul Sutcliffe, Jr., paul@devon.lns.pa.us,
@@ -310,7 +310,7 @@ static int lock_write_pid _P1((fd), int fd)
     if ( write(fd, &bpid, sizeof(bpid) ) != sizeof(bpid) )
 #else
     char apid[16];
-    sprintf( apid, "%10d\n", getpid() );
+    sprintf( apid, "%10d\n", (int) getpid() );
     if ( write(fd, apid, strlen(apid)) != strlen(apid) )
 #endif
     {
